Move JavaPathProvider to Rapicgen.Core.External namespace
